Add mobile navigation functionality
Implemented the display and behavior for mobile navigation. This includes adding a menu icon for mobile view and toggling the menu visibility on click. Adjustments were made to the styles and script to cater for these changes, as well as the incorporation of a dropdown menu that appears when the icon is clicked. Also, indentation in the CSS file has been uniformly fixed.