Update navigation menu for mobile view
The MainNavigation component was modified to better handle mobile view. Changes include adding a close icon and adjusting display settings to add modal behavior, which switches the menu between hidden and visible states when clicked. Also, the SVG for the close icon was updated and the media query limit was altered in MainHeader.vue to provide a more precise breakpoint.