Update footer styling and add link to logo
This commit modified the styling for the footer section of the website to improve its appearance on larger screens. Also, a RouterLink was added to the logo in the footer, which now redirects users to the home page when clicked. The changes aim to enhance user experience, especially for those on larger screen devices.